[QA] Move addStaticFile call to TCA overrides
Adding and manipulating the TCA inside the ext_tables.php and ext_localconf.php
has been fully deprecated in TYPO3 8.6.0-dev. Furthermore, the ext_tables.php is
not even being loaded when TYPO3 is in FE mode without logged in BE user anymore.

Instead, API calls that change TCA and direct changes of TCA should be moved to
the TCA overrides file for the changed table.

This way, the entire TCA can be cached once generated and does not have to be
manipulated over and over on each page render.